Add files via upload
This model encompasses unsupervised learning algorithms using Scikit library. 
It is to cluster customers in the behavior of online shopping. 
In this repository, libraries in Jupyter notebook is provided. Data processing of outliers is done to be within a variance range. 
Developed unsupervised algorithms include:
1- KMeans 
2- DBSCAN 
3- Meanshift 
4- MiniBatch 
5- Gaussian Mixture Model (GMM)

The following evaluations are performed: 
1- Elbow method for optimal number of clusters 
2- Silouette coefficient 
3- Calinski Harabasz 
4- Bayesian Information Criterion (BIC) for GMM algorithm 
5- GridSearch CV for model optimization 

Then, Facetgrid , GMM Log Likelihood, as well as covariance ellipses visualizations are included.